% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusrank.R
\docType{package}
\name{clusrank-package}
\alias{clusrank}
\alias{clusrank-package}
\title{Rank tests for clustered data.}
\description{
clusrank provides Wilcoxon rank-sum test and Wilcoxon sign-rank test for
clustered data. Each test has two methods: RGL and DS. The two tests were
unified with one single interface: \code{\link{clusWilcox.test}}.
}
\references{
Bernard Rosner, Robert J. Glynn, Mei-Ling T. Lee (2006)
\emph{The Wilcoxon Signed Rank Test for Paired Comparisons of
 Clustered Data}. Biometrics, \bold{62}, 185-192.

Bernard Rosner, Robert J. Glynn, Mei-Ling T. Lee (2003)
\emph{Incorporation of Clustering Effects for the Wilcoxon Rank
Sum Test: A Large-Sample Approach}. Biometrics, \bold{59}, 1089-1098.

Bernard Rosner, Robert J. Glynn, Mei-Ling T. Lee (2006)
\emph{Extension of the Rank Sum Test for Clustered Data:
Two-Group Comparisons with Group}. Biometrics, \bold{62}, 1251-1259.

Somnath Datta, Glen A. Satten (2005) \emph{Rank-Sum Tests for Clustered Data}.
Journal of the American Statistical Association, \bold{100}, 908-915.

Somath Datta, Glen A. Satten (2008) \emph{A Signed-Rank test for Clustered Data}.
Biometric, \bold{64}, 501-507.
}
\author{
\strong{Maintainer}: Yujing Jiang \email{yujing.jiang@uconn.edu}

Other contributors:
\itemize{
  \item Mei-Ling Ting Lee [contributor]
  \item Jun Yan [contributor]
}

}
