% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{cluster}
\alias{cluster}
\title{Identify clusters}
\usage{
cluster(x)
}
\arguments{
\item{x}{A numeric variable of cluster id.}
}
\value{
x
}
\description{
This is a special function used in the context of formula
used for Wilcoxon sum rank test for clustered data.
It identifies the cluster id of observations, and is used
on the right hand side of a formula.
}
\details{
THe function's only action is semantic, to mark
a variable as the cluster indicator. If not supplied,
will assume no cluster in the data.
}
\seealso{
\code{\link{clusWilcox.test.formula}}
}
\keyword{internal}
