% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crd.R
\docType{data}
\name{crd}
\alias{crd}
\title{Clustered non-stratified data}
\format{A data frame with 340 rows and 3 variables}
\source{
\url{https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/incorporating-cluster-effects-for-
the-wilcoxon-rank-sum-test-1/cluswilcox/sample_data}
}
\description{
non-stratified data include 3 variables:
 id, z, group. Have some missing values.
}
\details{
\itemize{
  \item id. cluster id
  \item z. observed score
  \item group. group id
}
}
\keyword{datasets}

