% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crdstr.R
\docType{data}
\name{crdStr}
\alias{crdStr}
\title{Clustered Stratified Data for Testing the Clustered Rank Sum Test}
\format{
A data frame with 956 observations on 4 variables.

[, 1] id      cluster id

[, 2] z       score

[, 3] group   group id

[, 4] stratum stratum id
}
\source{
\url{https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/incorporating-cluster-effects-for-the-wilcoxon-rank-sum-test-1/stratify--cluswilcox/sample-data}
}
\description{
Stratified data with clustering structure, Have some missing values.
}
\keyword{datasets}
\keyword{internal}
