% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsd.R
\docType{data}
\name{crsd}
\alias{crsd}
\title{Difference Between Pre and Post Treatment Scores with Clustering
 Structure: Balanced}
\format{
A data frame with 40 rows and 2 variables.


  [, 1] id cluster id

  [, 2] z  score difference
}
\source{
\url{https://sites.google.com/a/channing.
harvard.edu/bernardrosner/channing/cluster-signed
-rank-test-for-balancedunbalanced-design/sample_balanced_data}
}
\description{
Simulated clustered data with the difference of pre and post treatment scores recorded.
 The data is balanced.
}
\keyword{datasets}
\keyword{internal}
