% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsdunb.R
\docType{data}
\name{crsdUnb}
\alias{crsdUnb}
\title{Difference between Pre and Post Treatment Scores with Clustering
 Structure: Unbalanced}
\format{
A data frame with 748 rows and 2 variables.

  [, 1] id cluster id

  [, 2] z  score difference
}
\source{
\url{https://sites.google.com/a/channing.harvard.edu
/bernardrosner/channing/
cluster-signed-rank-test-for
-balancedunbalanced-design/sample-data-for-unbalanced-data-1}
}
\description{
Simulated clustered data with the difference of pre and post treatment scores recorded.
 The data is unbalanced.
}
\keyword{datasets}
\keyword{internal}
