% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{stratum}
\alias{stratum}
\title{Identify strata.}
\usage{
stratum(x)
}
\arguments{
\item{x}{A numeric variable of stratum id.}
}
\description{
This is a special function used in the context of formula
used for Wilcoxon sum rank test for clustered data.
It identifies the stratum id of observations, and is used
on the right hand side of a formula.
}
\details{
THe function's only action is semantic, to mark
a variable as the stratum indicator. If not supplied,
will assume no stratification in the data.
}
\seealso{
clusWilcox.test.formula
}
\keyword{internal}
