% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crd.R
\docType{data}
\name{crd}
\alias{crd}
\title{Clustered Non-Stratified Data for Testing Clustered Rank Sum Test}
\format{
A data frame with 340 observations on 3 variables.

[, 1] id    cluster id

[, 2] z     score

[, 3] group group id
}
\source{
\url{https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/incorporating-cluster-effects-for-
the-wilcoxon-rank-sum-test-1/cluswilcox/sample_data}
}
\description{
Non-stratified data including 3 variables:
 id, z, group. Have some missing values.
}
\keyword{datasets}
\keyword{internal}
