% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crdstr.R
\name{crdStr}
\alias{crdStr}
\title{Clustered stratified data}
\format{A data frame with 956 rows and 4 variables}
\source{
\url{https://sites.google.com/a/channing.harvard.edu/bernardrosner/channing/incorporating-cluster-effects-for-the-wilcoxon-rank-sum-test-1/stratify--cluswilcox/sample-data}
}
\description{
Stratified data for test with clustering structure
 id, z, group, stratum. Have some missing values.
}
\details{
\itemize{
  \item id. cluster id
  \item z. observed score
  \item group. group id
  \item stratum. stratum id
}
}

