% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsd.R
\name{crsd}
\alias{crsd}
\title{Difference between pre and post treatment scores with clustering
 structure.}
\format{A data frame with 40 rows and 2 variables}
\source{
\url{https://sites.google.com/a/channing.
harvard.edu/bernardrosner/channing/cluster-signed
-rank-test-for-balancedunbalanced-design/sample_balanced_data}
}
\description{
Balanced data include 2 variables:
 id, z.
}
\details{
\itemize{
  \item id. cluster id
  \item z. observed score
}
}

