% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crsdunb.R
\name{crsdUnb}
\alias{crsdUnb}
\title{Difference between pre and post treatment scores with clustering
 structure with variable cluster size.}
\format{A data frame with 748 rows and 2 variables}
\source{
\url{https://sites.google.com/a/channing.harvard.edu
/bernardrosner/channing/
cluster-signed-rank-test-for
-balancedunbalanced-design/sample-data-for-unbalanced-data-1}
}
\description{
Unbalanced data include 2 variables:
 id, z.
}
\details{
\itemize{
  \item id. cluster id
  \item z. observed score
}
}

